requirements-dev.txt
pytest==2.7.1
pytest-django==2.8.0
Sphinx==1.3.1
mock==1.0.1
tox==2.0.2
sphinxcontrib-napoleon==0.3.3
